Python version is being strange, moving the string up a level seems to fix it?
